test service changed n to 2.5 for fomrula algotrithm

+++ app/src/main/java/com/kiowa/bluetoothlocatability/utilities/Formulas.kt
@@ -13,7 +13,7 @@ class Formulas {
          * @author Kiowa Daly
          * @return Double
          */
-        fun rssiDistanceFormula(rssi: Double, txPower: Double, n: Int): Double {
+        fun rssiDistanceFormula(rssi: Double, txPower: Double, n: Double): Double {
             //d = 10 ^ ((txPower-Rssi) / 10n) (n ranges from 2 to 4)
             val bottom = (10.0 * n)
             val top = (txPower - (rssi))
